Refactor strategy package and timezone helpers
