+++ domain/models/log.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from dataclasses import dataclass
 from datetime import datetime
 
-from ._timezone import ensure_belgrade_timezone
+from ._timezone import ensure_current_timezone
 
 
 @dataclass(frozen=True, slots=True)
@@ -17,7 +17,7 @@ class LogLine:
     level: str
 
     def __post_init__(self) -> None:
-        ensure_belgrade_timezone(self.timestamp)
+        ensure_current_timezone(self.timestamp)
 
 
 __all__ = ["LogLine"]
