+++ domain/models/order_book.py
@@ -6,7 +6,7 @@ from dataclasses import dataclass
 from datetime import datetime
 from typing import Tuple
 
-from ._timezone import ensure_belgrade_timezone
+from ._timezone import ensure_current_timezone
 from .enums import Exchange
 
 
@@ -23,7 +23,7 @@ class OrderBookLevel:
     max_quantity_seen: float
 
     def __post_init__(self) -> None:
-        ensure_belgrade_timezone(self.first_seen_at, self.last_update_at)
+        ensure_current_timezone(self.first_seen_at, self.last_update_at)
 
 
 @dataclass(frozen=True, slots=True)
@@ -38,7 +38,7 @@ class OrderBookSnapshot:
     received_at: datetime
 
     def __post_init__(self) -> None:
-        ensure_belgrade_timezone(self.received_at)
+        ensure_current_timezone(self.received_at)
 
 
 @dataclass(frozen=True, slots=True)
@@ -54,7 +54,7 @@ class OrderBookUpdate:
     event_time: datetime
 
     def __post_init__(self) -> None:
-        ensure_belgrade_timezone(self.event_time)
+        ensure_current_timezone(self.event_time)
 
 
 __all__ = [
