+++ domain/models/position.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from dataclasses import dataclass
 from datetime import datetime
 
-from ._timezone import ensure_belgrade_timezone
+from ._timezone import ensure_current_timezone
 from .enums import Exchange, MarginMode, Side
 
 
@@ -25,7 +25,7 @@ class Position:
     updated_at: datetime
 
     def __post_init__(self) -> None:
-        ensure_belgrade_timezone(self.opened_at, self.updated_at)
+        ensure_current_timezone(self.opened_at, self.updated_at)
 
 
 __all__ = ["Position"]
