+++ domain/models/pressure.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from dataclasses import dataclass
 from datetime import datetime
 
-from ._timezone import ensure_belgrade_timezone
+from ._timezone import ensure_current_timezone
 
 
 @dataclass(frozen=True, slots=True)
@@ -18,7 +18,7 @@ class Pressure:
     imbalance_ratio: float
 
     def __post_init__(self) -> None:
-        ensure_belgrade_timezone(self.computed_at)
+        ensure_current_timezone(self.computed_at)
 
 
 __all__ = ["Pressure"]
