+++ domain/strategy/telegram.py
@@ -1,0 +1,31 @@
+"""Telegram notifications for the strategy."""
+
+from __future__ import annotations
+
+from dataclasses import dataclass
+from datetime import datetime
+
+from domain.models import Signal
+
+from .types import TelegramHandler
+
+
+@dataclass
+class TelegramNotifier:
+    """Sends strategy updates via Telegram."""
+
+    send_message: TelegramHandler
+
+    def notify_uptick(self, symbol: str, signal: Signal, timestamp: datetime) -> None:
+        direction = "лонг" if signal is Signal.LONG else "шорт"
+        self.send_message(f"{timestamp:%H:%M:%S} Обнаружен аптик {symbol} {direction}.")
+
+    def notify_entry(self, symbol: str, signal: Signal, timestamp: datetime) -> None:
+        direction = "лонг" if signal is Signal.LONG else "шорт"
+        self.send_message(f"{timestamp:%H:%M:%S} Открыт {direction} {symbol}.")
+
+    def notify_exit(self, symbol: str, reason: str, timestamp: datetime) -> None:
+        self.send_message(f"{timestamp:%H:%M:%S} Закрыт {symbol}. {reason}.")
+
+
+__all__ = ["TelegramNotifier"]
